+++ carla_ros_bridge/src/carla_ros_bridge/sensor.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-
 #
 # Copyright (c) 2018-2019 Intel Corporation
 #
@@ -10,12 +9,9 @@
 Classes to handle Carla sensors
 """
 from abc import abstractmethod
-
-import threading
+import queue
 
 import rospy
-import carla_ros_bridge.transforms as trans
-
 
 from carla_ros_bridge.actor import Actor
 
@@ -26,7 +22,16 @@ class Sensor(Actor):
     Actor implementation details for sensors
     """
 
-    def __init__(self, carla_actor, parent, communication, prefix=None):
+    def __init__(self,  # pylint: disable=too-many-arguments
+                 carla_actor,
+                 parent,
+                 communication,
+                 synchronous_mode,
+                 is_event_sensor=False,  # only relevant in synchronous_mode:
+                 # if a sensor only delivers data on special events,
+                 # do not wait for it. That means you might get data from a
+                 # sensor, that belongs to a different frame
+                 prefix=None):
         """
         Constructor
 
@@ -36,6 +41,8 @@ class Sensor(Actor):
         :type parent: carla_ros_bridge.Parent
         :param communication: communication-handle
         :type communication: carla_ros_bridge.communication
+        :param synchronous_mode: use in synchronous mode?
+        :type synchronous_mode: bool
         :param prefix: the topic prefix to be used for this actor
         :type prefix: string
         """
@@ -46,8 +53,17 @@ class Sensor(Actor):
                                      communication=communication,
                                      prefix=prefix)
 
-        self.current_sensor_data = None
-        self.update_lock = threading.Lock()
+        self.synchronous_mode = synchronous_mode
+        self.queue = queue.Queue()
+        self.next_data_expected_time = None
+        self.sensor_tick_time = None
+        self.is_event_sensor = is_event_sensor
+        try:
+            self.sensor_tick_time = float(carla_actor.attributes["sensor_tick"])
+            rospy.logdebug("Sensor tick time is {}".format(self.sensor_tick_time))
+        except (KeyError, ValueError):
+            self.sensor_tick_time = None
+
         if self.__class__.__name__ != "Sensor":
             self.carla_actor.listen(self._callback_sensor_data)
 
@@ -63,8 +79,6 @@ class Sensor(Actor):
         rospy.logdebug("Destroy Sensor(id={})".format(self.get_id()))
         if self.carla_actor.is_listening:
             self.carla_actor.stop()
-        if self.update_lock.acquire():
-            self.current_sensor_data = None
         super(Sensor, self).destroy()
 
     def _callback_sensor_data(self, carla_sensor_data):
@@ -75,11 +89,14 @@ class Sensor(Actor):
         :type carla_sensor_data: carla.SensorData
         """
         if not rospy.is_shutdown():
-            if self.update_lock.acquire(False):
-                self.current_sensor_data = carla_sensor_data
-                self.publish_transform()
+            if self.synchronous_mode:
+                if self.sensor_tick_time:
+                    self.next_data_expected_time = carla_sensor_data.timestamp + \
+                        float(self.sensor_tick_time)
+                self.queue.put(carla_sensor_data)
+            else:
+                self.publish_transform(self.get_ros_sensor_transform(carla_sensor_data.transform))
                 self.sensor_data_updated(carla_sensor_data)
-                self.update_lock.release()
 
     @abstractmethod
     def sensor_data_updated(self, carla_sensor_data):
@@ -93,10 +110,67 @@ class Sensor(Actor):
         raise NotImplementedError(
             "This function has to be implemented by the derived classes")
 
-    def get_current_ros_transform(self):
+    def get_ros_sensor_transform(self, transform):
+        """
+        Get sensor tf (override, if required)
 
-        tf_msg = super(Sensor, self).get_current_ros_transform()
-        # sensor data has its own transform
-        tf_msg.transform = trans.carla_transform_to_ros_transform(
-            self.current_sensor_data.transform)
+        :param transform: carla sensor transform
+        :type transform: carla.Transform
+        """
+        tf_msg = super(Sensor, self).get_ros_transform(transform)
         return tf_msg
+
+    def _update_synchronous_event_sensor(self, frame, timestamp):
+        while True:
+            try:
+                carla_sensor_data = self.queue.get(block=False)
+                if carla_sensor_data.frame != frame:
+                    rospy.logwarn("{}({}): Received event for frame {}"
+                        " (expected {}). Process it anyways.".format(
+                        self.__class__.__name__,
+                        self.get_id(),
+                        carla_sensor_data.frame,
+                        frame))
+                rospy.loginfo("{}({}): process {}".format(
+                    self.__class__.__name__, self.get_id(), frame))
+                self.publish_transform(
+                    self.get_ros_transform(carla_sensor_data.transform))
+                self.sensor_data_updated(carla_sensor_data)
+            except queue.Empty:
+                return
+        
+    def _update_synchronous_sensor(self, frame, timestamp):
+        while not self.next_data_expected_time or\
+            (not self.queue.empty() or
+             self.next_data_expected_time and
+             self.next_data_expected_time < timestamp):
+            while True:
+                try:
+                    carla_sensor_data = self.queue.get(timeout=1.0)
+                    if carla_sensor_data.frame == frame:
+                        rospy.loginfo("{}({}): process {}".format(
+                            self.__class__.__name__, self.get_id(), frame))
+                        self.publish_transform(
+                            self.get_ros_transform(carla_sensor_data.transform))
+                        self.sensor_data_updated(carla_sensor_data)
+                        return
+                    else:
+                        rospy.logwarn("{}({}): skipping old frame {}, expected {}".format(
+                            self.__class__.__name__,
+                            self.get_id(),
+                            carla_sensor_data.frame,
+                            frame))
+                except queue.Empty:
+                    if not rospy.is_shutdown():
+                        rospy.logwarn("{}({}): Expected Frame {} not received".format(
+                            self.__class__.__name__, self.get_id(), frame))
+                    return
+
+    def update(self, frame, timestamp):
+        if self.synchronous_mode:
+            if self.is_event_sensor:
+                self._update_synchronous_event_sensor(frame, timestamp)
+            else:
+                self._update_synchronous_sensor(frame, timestamp)
+                
+        super(Sensor, self).update(frame, timestamp)